built-in/requirements.txt
# Sign Language Reader - Python Dependencies

# Core Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0

# GUI
pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0

# MQTT Communication
paho-mqtt>=1.6.0
